+++ venice/api/ticker.py
@@ -1,12 +1,15 @@
+from decimal import Decimal
+
+
 class Ticker:
     def __init__(self, time, ask, bid, last, low=0, high=0, volume=0):
         self.time = time
-        self.ask = float(ask)
-        self.bid = float(bid)
-        self.last = float(last)
-        self.low = float(low)
-        self.high = float(high)
-        self.volume = float(volume)
+        self.ask = Decimal(ask)
+        self.bid = Decimal(bid)
+        self.last = Decimal(last)
+        self.low = Decimal(low)
+        self.high = Decimal(high)
+        self.volume = Decimal(volume)
 
     def __str__(self):
         return '{} ask:{}, bid:{}, last:{}, low:{}, high:{}, vol:{}'.format(
